Moves the one of merge all patient db logic into the mgmt command
